Make bookinfo accessible for the tracing env.
