+++ apps/bookinfo/lib.rs
@@ -1,0 +1,6 @@
+mod gateway;
+mod leafnode;
+mod productpage;
+mod reviews;
+
+pub mod bookinfo;
